+++ modules/asset/exported/exported.go
@@ -1,7 +1,7 @@
 package exported
 
 import (
-	"github.com/irisnet/irishub/modules/asset/types"
+	"github.com/irisnet/irishub/modules/asset/internal/types"
 )
 
 type FungibleToken = types.FungibleToken
