+++ modules/asset/internal/keeper/fees.go
@@ -6,7 +6,7 @@ import (
 	"strconv"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/irisnet/irishub/modules/asset/types"
+	"github.com/irisnet/irishub/modules/asset/internal/types"
 )
 
 // fee factor formula: (ln(len({name}))/ln{base})^{exp}
