+++ modules/asset/internal/keeper/querier.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/irisnet/irishub/modules/asset/types"
+	"github.com/irisnet/irishub/modules/asset/internal/types"
 	iristypes "github.com/irisnet/irishub/types"
 	abci "github.com/tendermint/tendermint/abci/types"
 )
