+++ modules/asset/simulation/genesis.go
@@ -4,12 +4,12 @@ package simulation
 
 import (
 	"fmt"
-	"github.com/irisnet/irishub/modules/asset/types"
 	"math/rand"
 
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/module"
+	"github.com/irisnet/irishub/modules/asset/internal/types"
 )
 
 // Simulation parameter constants
